Major refactor.
Move large text pieces to constants.
Rename handler.py and split each update type into separate classes.
